+++ android/src/main/java/slayer/accessibility/service/flutter_accessibility_service/FlutterAccessibilityServicePlugin.java
@@ -1,5 +1,6 @@
 package slayer.accessibility.service.flutter_accessibility_service;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
@@ -9,16 +10,19 @@ import android.util.Log;
 import androidx.annotation.NonNull;
 
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
+import io.flutter.embedding.engine.plugins.activity.ActivityAware;
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 import io.flutter.plugin.common.EventChannel;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
 import io.flutter.plugin.common.MethodChannel.Result;
+import io.flutter.plugin.common.PluginRegistry;
 
 /**
  * FlutterAccessibilityServicePlugin
  */
-public class FlutterAccessibilityServicePlugin implements FlutterPlugin, MethodCallHandler, EventChannel.StreamHandler {
+public class FlutterAccessibilityServicePlugin implements FlutterPlugin, ActivityAware, MethodCallHandler, PluginRegistry.ActivityResultListener, EventChannel.StreamHandler {
 
     private static final String CHANNEL_TAG = "x-slayer/accessibility_channel";
     private static final String EVENT_TAG = "x-slayer/accessibility_event";
@@ -26,6 +30,10 @@ public class FlutterAccessibilityServicePlugin implements FlutterPlugin, MethodC
     private MethodChannel channel;
     private EventChannel eventChannel;
     private Context context;
+    private Activity mActivity;
+
+    private Result pendingResult;
+    final int REQUEST_CODE_FOR_ACCESSIBILITY = 167;
 
     @Override
     public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
@@ -41,13 +49,12 @@ public class FlutterAccessibilityServicePlugin implements FlutterPlugin, MethodC
 
     @Override
     public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
+        pendingResult = result;
         if (call.method.equals("isAccessibilityPermissionEnabled")) {
             result.success(Utils.isAccessibilitySettingsOn(context));
         } else if (call.method.equals("requestAccessibilityPermission")) {
             Intent intent = new Intent(Settings.ACTION_ACCESSIBILITY_SETTINGS);
-            intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
-            context.startActivity(intent);
-            result.success(null);
+            mActivity.startActivityForResult(intent, REQUEST_CODE_FOR_ACCESSIBILITY);
         } else {
             result.notImplemented();
         }
@@ -79,5 +86,41 @@ public class FlutterAccessibilityServicePlugin implements FlutterPlugin, MethodC
     @Override
     public void onCancel(Object arguments) {
         eventChannel.setStreamHandler(null);
+    }
+
+    @Override
+    public boolean onActivityResult(int requestCode, int resultCode, Intent data) {
+        if (requestCode == REQUEST_CODE_FOR_ACCESSIBILITY) {
+            if (resultCode == Activity.RESULT_OK) {
+                pendingResult.success(true);
+            } else if (resultCode == Activity.RESULT_CANCELED) {
+                pendingResult.success(Utils.isAccessibilitySettingsOn(context));
+            } else {
+                pendingResult.success(false);
+            }
+            return true;
+        }
+        return false;
+    }
+
+    @Override
+    public void onAttachedToActivity(@NonNull ActivityPluginBinding binding) {
+        this.mActivity = binding.getActivity();
+        binding.addActivityResultListener(this);
+    }
+
+    @Override
+    public void onDetachedFromActivityForConfigChanges() {
+        this.mActivity = null;
+    }
+
+    @Override
+    public void onReattachedToActivityForConfigChanges(@NonNull ActivityPluginBinding binding) {
+        onAttachedToActivity(binding);
+    }
+
+    @Override
+    public void onDetachedFromActivity() {
+        this.mActivity = null;
     }
 }